Change type: just float accepted

+++ codigo.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 #include <math.h>
 void teste(float a) {
-printf("%d\n", a);
+printf("%f\n", a);
 }
 int main() {
-int a;
-int b;
+float a;
+float b;
 a = 2.0  ;
 b = 1.0  ;
 printf("%s %f\n", "OLas", a);
